+++ modules/task_2/shagov_m_fox/fox.h
@@ -13,7 +13,7 @@ typedef  std::vector<double> Matrix;
 
 bool isEqual(double x, double y);
 bool isEqualMatrix(const Matrix& A, const Matrix& B);
-bool isSizeCorrect(size_t size, int t_count);
+bool isSizeCorrect(size_t size, size_t t_count);
 
 Matrix createRandomMatrix(size_t size);
 Matrix sequentialMatrixMultiplication(const std::vector<double>& A, const std::vector<double>& B, size_t BlockSize);
